#903: trace: Extract code used in MPI traces to separate struct TraceLite
