+++ src/vt/trace/trace.h
@@ -49,7 +49,7 @@
 #include "vt/trace/trace_containers.h"
 #include "vt/trace/trace_log.h"
 #include "vt/trace/trace_registry.h"
-#include "vt/trace/trace_user_event.h"
+#include "vt/trace/trace_lite.h"
 #include "vt/runtime/component/component_pack.h"
 
 #include "vt/timing/timing.h"
@@ -64,8 +64,6 @@
 #include <queue>
 
 namespace vt { namespace trace {
-
-struct vt_gzFile;
 
 /// Tracking information for beginProcessing/endProcessing.
 struct TraceProcessingTag {
@@ -109,15 +107,8 @@ private:
  * Outputs Projections log and sts files, which can be examined by the
  * java Projections tool.
  */
-struct Trace : runtime::component::Component<Trace> {
-  using LogType             = Log;
-  using TraceConstantsType  = eTraceConstants;
-  using TimeIntegerType     = int64_t;
-  using TraceContainerType  = std::queue<LogType>;
-  // Although should be used mostly as a stack, vector is exposed to enable
-  // the use of a synthetic pop-push to maintain the stack around idle.
-  using TraceStackType      = std::vector<LogType>;
-  using EventHoldStackType  = std::vector<std::size_t>;
+struct Trace : runtime::component::Component<Trace>, TraceLite {
+
 
   /**
    * \internal \brief System call to construct the trace component
@@ -132,52 +123,9 @@ struct Trace : runtime::component::Component<Trace> {
 
   friend struct Log;
 
-  /**
-   * \brief Get the trace file name for this node
-   *
-   * \return the file name
-   */
-  std::string getTraceName() const { return full_trace_name_; }
-
-  /**
-   * \brief Get the sts file name that holds meta-data
-   *
-   * \return the sts file name
-   */
-  std::string getSTSName()   const { return full_sts_name_;   }
-
-  /**
-   * \brief Get the trace directory for output
-   *
-   * \return the directory path
-   */
-  std::string getDirectory() const { return full_dir_name_;   }
-
   void initialize() override;
   void startup() override;
   void finalize() override;
-
-  /**
-   * \brief Initialize trace module in stand-alone mode. This will create
-   * stripped-down version of runtime and initialize components needed for
-   * tracing MPI calls
-   *
-   * \param[in] comm MPI communicator type
-   * \param[in] flush_size Flush output trace every (flush_size) trace records
-   */
-  void initializeStandalone(MPI_Comm comm, int32_t flush_size);
-
-  /**
-   * \brief Cleanup all components initialized for standalone mode
-   */
-  void finalizeStandalone();
-
-  /**
-   * \internal \brief Setup the file names for output
-   *
-   * \param[in] in_prog_name the program name
-   */
-  void setupNames(std::string const& in_prog_name);
 
   /**
    * \brief Initiate a paired processing event.
@@ -234,32 +182,6 @@ struct Trace : runtime::component::Component<Trace> {
   void endSchedulerLoop();
 
   /**
-   * \brief Scheduler trigger for \c sched::SchedulerEvent::BeginIdle
-   *
-   * \param[in] time time it begins idle
-   */
-  void beginIdle(double const time = getCurrentTime());
-
-  /**
-   * \brief Scheduler trigger for \c sched::SchedulerEvent::EndIdle
-   *
-   * \param[in] time time it ends idle
-   */
-  void endIdle(double const time = getCurrentTime());
-
-  /**
-   * \brief Collectively register a user event
-   *
-   * \note For users, it is recommended that the free function be called
-   * \c registerEventCollective
-   *
-   * \param[in] name name for the user event
-   *
-   * \return the user event ID
-   */
-  UserEventIDType registerUserEventColl(std::string const& name);
-
-  /**
    * \brief Register a user event rooted on a single node
    *
    * \note For users, it is recommended that the free function be called
@@ -314,15 +236,6 @@ struct Trace : runtime::component::Component<Trace> {
   void addUserEventManual(UserSpecEventIDType event);
 
   /**
-   * \brief Log a bracketed user event with start and end time
-   *
-   * \param[in] event the ID for the sts file
-   * \param[in] begin the begin time
-   * \param[in] end the end time
-   */
-  void addUserEventBracketed(UserEventIDType event, double begin, double end);
-
-  /**
    * \brief Log a bracketed user event manually with start and end time
    *
    * \param[in] event the ID for the sts file
@@ -374,19 +287,6 @@ struct Trace : runtime::component::Component<Trace> {
    * \param[in] data the integer to add
    */
   void addUserData(int32_t data);
-
-  /**
-   * \brief Log a user bracketed event with a note
-   *
-   * \param[in] begin the begin time
-   * \param[in] end the end time
-   * \param[in] note the note to log
-   * \param[in] event the event ID
-   */
-  void addUserBracketedNote(
-    double const begin, double const end, std::string const& note,
-    TraceEventIDType const event = no_trace_event
-  );
 
   /**
    * \brief Log a memory usage event
@@ -443,29 +343,6 @@ struct Trace : runtime::component::Component<Trace> {
   );
 
   /**
-   * \brief Enable logging of events at runtime
-   */
-  void enableTracing();
-
-  /**
-   * \brief Disable logging of events.
-   *
-   * \note Events already logged may still be written to the trace log.
-   */
-  void disableTracing();
-
-  /**
-   * \internal \brief Check if tracing is enabled
-   *
-   * \param[in] is_end_event whether the event that is being considering to
-   * write out in the calling context is actually an end event that needs to be
-   * closed
-   *
-   * \return whether tracing is enabled
-   */
-  bool checkDynamicRuntimeEnabled(bool is_end_event = false);
-
-  /**
    * \internal \brief Load and broadcast the trace specification file
    */
   void loadAndBroadcastSpec();
@@ -479,44 +356,11 @@ struct Trace : runtime::component::Component<Trace> {
   void setTraceEnabledCurrentPhase(PhaseType cur_phase);
 
   /**
-   * \brief Flush traces to file
-   *
-   * \param[in] useGlobalSync whether a global sync should be invoked before
-   * flushing output
-   */
-  void flushTracesFile(bool useGlobalSync);
-
-  /**
-   * \internal \brief Cleanup traces data, write to disk and close
-   */
-  void cleanupTracesFile();
-
-  /**
    * \brief Check if trace is in a idle event
    *
    * \return whether in an idle eveent
    */
   bool inIdleEvent() const;
-
-  /**
-   * \brief Get the current time
-   *
-   * \return query the current clock time
-   */
-  static inline double getCurrentTime() {
-    return ::vt::timing::Timing::getCurrentTime();
-  }
-
-  /**
-   * \brief Convert time in seconds to integer in microseconds
-   *
-   * \param[in] time the time in seconds as double
-   *
-   * \return time in microsecond as integer
-   */
-  static inline TimeIntegerType timeToMicros(double const time) {
-    return static_cast<TimeIntegerType>(time * 1e6);
-  }
 
   friend void insertNewUserEvent(UserEventIDType event, std::string const& name);
 
@@ -546,107 +390,6 @@ struct Trace : runtime::component::Component<Trace> {
   }
 
 private:
-  /**
-   * \brief Emit a 'stop' trace for previous open event or a '[re]start' trace
-   * for a reactivated open event. This assists with output flattening.
-   *
-   * \param[in] time the time
-   * \param[in] type type of event to emit
-   */
-  void emitTraceForTopProcessingEvent(
-    double const time, TraceConstantsType const type
-  );
-
-  /**
-   * \brief Writes traces to file, optionally flushing. The traces collection is
-   * modified.
-   *
-   * \param[in] file the gzip file to write to
-   * \param[in] traces the container of collected traces
-   * \param[in] start_time the start time
-   * \param[in] flush the flush mode
-   */
-  static void outputTraces(
-    vt_gzFile* file, TraceContainerType& traces,
-    double start_time, int flush
-  );
-
-  /**
-   * \brief Output the tracing header
-   *
-   * \param[in] file the gzip file
-   * \param[in] node the node outputting on
-   * \param[in] start the start time
-   */
-  static void outputHeader(
-    vt_gzFile* file, NodeType const node, double const start
-  );
-
-  /**
-   * \brief Output the tracing footer
-   *
-   * \param[in] file the gzip file
-   * \param[in] node the node outputting on
-   * \param[in] start the start time
-   */
-  static void outputFooter(
-    vt_gzFile* file, NodeType const node, double const start
-  );
-
-  /**
-   * \brief Write traces to file
-   *
-   * \param[in] flush the flush mode
-   * \param[in] is_incremental_flush whether this is an incremental flush
-   */
-  void writeTracesFile(int flush, bool is_incremental_flush);
-
-  /**
-   * \brief Output the sts (control) file
-   *
-   * \param[in] file the file
-   */
-  void outputControlFile(std::ofstream& file);
-
-  /**
-   * \brief Check if tracing is enabled on a certain node
-   *
-   * \param[in] node the node
-   *
-   * \return whether it is enabled
-   */
-  static bool traceWritingEnabled(NodeType node);
-
-  /**
-   * \brief Check if a node will output the sts file
-   *
-   * \param[in] node the node
-   *
-   * \return whether it will output
-   */
-  static bool isStsOutputNode(NodeType node);
-
-  /**
-   * \brief Log an event, returning a trace event ID if accepted
-   * or \c no_trace_event if not accepted (eg. no tracing on node).
-   * The log object is invalidated after the call.
-   *
-   * \param[in] log the entity to log
-   *
-   * \return the trace event ID for that new log
-   */
-  TraceEventIDType logEvent(LogType&& log);
-
-  /**
-   * \brief Get the current traces size
-   *
-   * \return computed bytes used for tracing (lower bound)
-   */
-  std::size_t getTracesSize() const {
-    return traces_.size() * sizeof(Log);
-  }
-
-private:
   /*
    * Incremental flush mode for zlib. Not set here with zlib constants to reduce
    * header dependencies.
@@ -654,29 +397,11 @@ private:
   int incremental_flush_mode = 0;
 
 private:
-  TraceContainerType traces_;
-  TraceStackType open_events_;
-  EventHoldStackType event_holds_;
 
-  TraceEventIDType cur_event_   = 1;
-  bool enabled_                 = true;
-  bool idle_begun_              = false;
-  double start_time_            = 0.0;
-  UserEventRegistry user_event_ = {};
-
-  std::string prog_name_        = "";
-  std::string trace_name_       = "";
-  std::string full_trace_name_  = "";
-  std::string full_sts_name_    = "";
-  std::string full_dir_name_    = "";
-  std::unique_ptr<vt_gzFile> log_file_;
-  bool wrote_sts_file_          = false;
-  size_t trace_write_count_     = 0;
-  bool standalone_version_      = false;
 
   ObjGroupProxyType spec_proxy_ = vt::no_obj_group;
-  bool trace_enabled_cur_phase_ = true;
-  UserEventIDType flush_event_  = no_user_event_id;
+
+
 
   // Processing event between top-level loops.
   TraceEntryIDType between_sched_event_type_ = no_trace_entry_id;
@@ -684,13 +409,5 @@ private:
 };
 
 }} //end namespace vt::trace
-
-namespace vt {
-
-#if vt_check_enabled(trace_enabled)
-  extern trace::Trace* theTrace();
-#endif
-
-}
 
 #endif /*INCLUDED_TRACE_TRACE_H*/
